+++ iModelCore/BeSQLite/PublicAPI/BeSQLite/CloudSqlite.h
@@ -51,7 +51,7 @@ struct CloudCache {
     Utf8String m_name;
     std::vector<CloudContainer*> m_containers;
 
-    BE_SQLITE_EXPORT CloudResult InitCache(Utf8StringCR name, Utf8StringCR rootDir, int64_t cacheSize = 0, int64_t nRequest = 0, int64_t httpTimeout = 0, bool verboseCurlLog = false);
+    BE_SQLITE_EXPORT CloudResult InitCache(Utf8StringCR name, Utf8StringCR rootDir, int64_t cacheSize = 0, int nRequest = 0, int httpTimeout = 0, bool verboseCurlLog = false);
     BE_SQLITE_EXPORT void SetLogMask(int logMask);
     BE_SQLITE_EXPORT Utf8String GetDatabaseHash(CloudContainer& container, Utf8StringCR dbName);
     BE_SQLITE_EXPORT int GetNumCleanupBlocks(CloudContainer& container);
